added eslint and typescript
