+++ package.json
@@ -11,11 +11,16 @@
     "@testing-library/jest-dom": "^5.11.4",
     "@testing-library/react": "^11.1.0",
     "@testing-library/user-event": "^12.1.10",
+    "@types/jest": "^27.4.0",
+    "@types/node": "^17.0.15",
+    "@types/react": "^17.0.39",
+    "@types/react-dom": "^17.0.11",
     "dompurify": "^2.3.5",
     "prop-types": "^15.7.2",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-scripts": "^4.0.3",
+    "typescript": "^4.5.5",
     "web-vitals": "^1.0.1"
   },
   "homepage": "https://nadiia-kobylinska.github.io/nk-notes",
@@ -23,7 +28,9 @@
     "start": "react-scripts start",
     "build": "react-scripts build",
     "test": "react-scripts test",
-    "eject": "react-scripts eject"
+    "eject": "react-scripts eject",
+    "lint": "eslint src/**/*.js src/**/*.jsx src/**/*.ts src/**/*.tsx",
+    "lint:fix": "eslint --fix src/**/*.js src/**/*.jsx src/**/*.ts src/**/*.tsx"
   },
   "eslintConfig": {
     "extends": [
@@ -33,7 +40,7 @@
   },
   "browserslist": {
     "production": [
-      ">0.2%",
+      ">5%",
       "not dead",
       "not op_mini all"
     ],
@@ -42,5 +49,10 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "5.11.0",
+    "@typescript-eslint/parser": "5.11.0",
+    "eslint-config-airbnb": "19.0.4"
   }
 }
