+++ src/components/notes-manage/form-view.js
@@ -4,8 +4,8 @@ import LoadingButton from "@mui/lab/LoadingButton";
 import CheckIcon from "@mui/icons-material/Check";
 import {useEffect, useReducer} from "react";
 import EditDiv from "../edit-div";
-import calcCountChar from "../../utility/calc-count-char";
-import cleanUpHTML from "../../utility/clean-up-html";
+import calcCountChar from "../../utils/calc-count-char";
+import cleanUpHTML from "../../utils/clean-up-html";
 import formReducer from "./formReducer";
 import RoundCounter from "../round-counter";
 import {isValid} from "./func";
