--- src/utility/clean-up-html.js
@@ -1,6 +1,0 @@
-import DOMPurify from 'dompurify'
-
-function cleanUpHTML(string){
-    return DOMPurify.sanitize(string,{ USE_PROFILES: { html: true } });
-}
-export default cleanUpHTML;
