+++ src/utils/clean-up-html.js
@@ -1,0 +1,6 @@
+import DOMPurify from 'dompurify';
+
+function cleanUpHTML(string) {
+  return DOMPurify.sanitize(string, { USE_PROFILES: { html: true } });
+}
+export default cleanUpHTML;
